@@ -5,14 +5,14 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.wj.ktolin.baselibrary"
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         ndk {
             abiFilters "armeabi", "armeabi-v7a", "x86", "mips"
         }
@@ -23,11 +23,11 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-        sourceSets {
-            main{
-                jniLibs.srcDirs = ['libs']
-            }
+    sourceSets {
+        main {
+            jniLibs.srcDirs = ['libs']
         }
+    }
     //指定jdk版本
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -37,14 +37,11 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation('com.android.support:appcompat-v7:28.0.0'){
-        exclude group: 'com.android.support', module: 'appcompat-v7'//从模块中删除
-    }
+//    implementation('androidx.appcompat:appcompat:1.0.0'){
+//        exclude group: 'com.android.support', module: 'appcompat-v7'//从模块中删除
+//    }
+    implementation 'androidx.appcompat:appcompat:1.0.2'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-//    implementation 'com.github.CarGuo:RecordWave:v1.1.8'
     implementation project(':ktlibrary')
-    //    implementation 'com.wj.libs:ktlibrary:1.0.0'
-    implementation('com.android.support:recyclerview-v7:28.0.0') {
-        exclude group: 'com.android.support', module: 'recyclerview-v7'//从模块中删除
-    }
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
 }
