@@ -15,7 +15,7 @@
  */
 package com.abase.okhttp.log;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.io.IOException;
 import java.net.InetAddress;
