@@ -10,7 +10,7 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.net.http.SslError;
 import android.os.Build;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.RequiresApi;
 import android.util.AttributeSet;
 import android.webkit.DownloadListener;
 import android.webkit.GeolocationPermissions;
