@@ -1,6 +1,6 @@
 package com.wj.ktolin.baselibrary;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewGroup;
 
