@@ -24,7 +24,7 @@ import android.content.ActivityNotFoundException
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 
 
 inline fun <reified T: Activity> Context.startActivity(vararg params: Pair<String, Any?>) =
@@ -33,13 +33,13 @@ inline fun <reified T: Activity> Context.startActivity(vararg params: Pair<Strin
 inline fun <reified T: Activity> AnkoContext<*>.startActivity(vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStartActivity(ctx, T::class.java, params)
 
-inline fun <reified T: Activity> Fragment.startActivity(vararg params: Pair<String, Any?>) =
+inline fun <reified T: Activity> androidx.fragment.app.Fragment.startActivity(vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStartActivity(this.activity!!, T::class.java, params)
 
 inline fun <reified T: Activity> Activity.startActivityForResult(requestCode: Int, vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStartActivityForResult(this, T::class.java, requestCode, params)
 
-inline fun <reified T: Activity> Fragment.startActivityForResult(requestCode: Int, vararg params: Pair<String, Any?>) =
+inline fun <reified T: Activity> androidx.fragment.app.Fragment.startActivityForResult(requestCode: Int, vararg params: Pair<String, Any?>) =
         startActivityForResult(AnkoInternals.createIntent(this.activity!!, T::class.java, params), requestCode)
 
 inline fun <reified T: Service> Context.startService(vararg params: Pair<String, Any?>) =
@@ -48,7 +48,7 @@ inline fun <reified T: Service> Context.startService(vararg params: Pair<String,
 inline fun <reified T: Service> AnkoContext<*>.startService(vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStartService(ctx, T::class.java, params)
 
-inline fun <reified T: Service> Fragment.startService(vararg params: Pair<String, Any?>) =
+inline fun <reified T: Service> androidx.fragment.app.Fragment.startService(vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStartService(this.activity!!, T::class.java, params)
 
 inline fun <reified T : Service> Context.stopService(vararg params: Pair<String, Any?>) =
@@ -57,7 +57,7 @@ inline fun <reified T : Service> Context.stopService(vararg params: Pair<String,
 inline fun <reified T : Service> AnkoContext<*>.stopService(vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStopService(ctx, T::class.java, params)
 
-inline fun <reified T : Service> Fragment.stopService(vararg params: Pair<String, Any?>) =
+inline fun <reified T : Service> androidx.fragment.app.Fragment.stopService(vararg params: Pair<String, Any?>) =
         AnkoInternals.internalStopService(this.activity!!, T::class.java, params)
 
 inline fun <reified T: Any> Context.intentFor(vararg params: Pair<String, Any?>): Intent =
@@ -66,7 +66,7 @@ inline fun <reified T: Any> Context.intentFor(vararg params: Pair<String, Any?>)
 inline fun <reified T: Any> AnkoContext<*>.intentFor(vararg params: Pair<String, Any?>): Intent =
         AnkoInternals.createIntent(ctx, T::class.java, params)
 
-inline fun <reified T: Any> Fragment.intentFor(vararg params: Pair<String, Any?>): Intent =
+inline fun <reified T: Any> androidx.fragment.app.Fragment.intentFor(vararg params: Pair<String, Any?>): Intent =
         AnkoInternals.createIntent(this.activity!!, T::class.java, params)
 
 /**
@@ -133,7 +133,7 @@ inline fun Intent.noHistory(): Intent = apply { addFlags(Intent.FLAG_ACTIVITY_NO
 inline fun Intent.singleTop(): Intent = apply { addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP) }
 
 inline fun AnkoContext<*>.browse(url: String, newTask: Boolean = false) = ctx.browse(url, newTask)
-inline fun Fragment.browse(url: String, newTask: Boolean = false) = activity?.browse(url, newTask)
+inline fun androidx.fragment.app.Fragment.browse(url: String, newTask: Boolean = false) = activity?.browse(url, newTask)
 
 fun Context.browse(url: String, newTask: Boolean = false): Boolean {
     try {
@@ -151,7 +151,7 @@ fun Context.browse(url: String, newTask: Boolean = false): Boolean {
 }
 
 inline fun AnkoContext<*>.share(text: String, subject: String = "") = ctx.share(text, subject)
-inline fun Fragment.share(text: String, subject: String = "") = activity?.share(text, subject)
+inline fun androidx.fragment.app.Fragment.share(text: String, subject: String = "") = activity?.share(text, subject)
 
 fun Context.share(text: String, subject: String = ""): Boolean {
     try {
@@ -168,7 +168,7 @@ fun Context.share(text: String, subject: String = ""): Boolean {
 }
 
 inline fun AnkoContext<*>.email(email: String, subject: String = "", text: String = "") = ctx.email(email, subject, text)
-inline fun Fragment.email(email: String, subject: String = "", text: String = "") = activity?.email(email, subject, text)
+inline fun androidx.fragment.app.Fragment.email(email: String, subject: String = "", text: String = "") = activity?.email(email, subject, text)
 
 fun Context.email(email: String, subject: String = "", text: String = ""): Boolean {
     val intent = Intent(Intent.ACTION_SENDTO)
@@ -187,7 +187,7 @@ fun Context.email(email: String, subject: String = "", text: String = ""): Boole
 }
 
 inline fun AnkoContext<*>.makeCall(number: String): Boolean = ctx.makeCall(number)
-inline fun Fragment.makeCall(number: String): Boolean = activity?.makeCall(number)!!
+inline fun androidx.fragment.app.Fragment.makeCall(number: String): Boolean = activity?.makeCall(number)!!
 
 @SuppressLint("MissingPermission")
 fun Context.makeCall(number: String): Boolean {
@@ -202,7 +202,7 @@ fun Context.makeCall(number: String): Boolean {
 }
 
 inline fun AnkoContext<*>.sendSMS(number: String, text: String = ""): Boolean = ctx.sendSMS(number, text)
-inline fun Fragment.sendSMS(number: String, text: String = ""): Boolean = activity?.sendSMS(number, text)!!
+inline fun androidx.fragment.app.Fragment.sendSMS(number: String, text: String = ""): Boolean = activity?.sendSMS(number, text)!!
 
 fun Context.sendSMS(number: String, text: String = ""): Boolean {
     try {
